Conform serialization to ngen BMI serialization protocol
Fixing conformance with ngen protocol:

- add include/ngen_utilities.h holding the reserved name, type, unit and
  item count tables, indexed by a shared enum
- rename involved variables to have `ngen::` prefix
- resolve the variables in Get_var_units, which is the engine's entire
  support probe and previously could never find them
- tweak types of serialization_size and serialization_create to int
- keep the variables out of output_var_names and input_var_names, and
  leave Get_var_grid and Get_var_location failing for them
- don't fail Get_var_nbytes or Get_value_ptr during restore, when they
  are needed but no buffer exists yet

+++ include/topmodel.h
@@ -6,7 +6,6 @@
 #include <float.h>
 #include <math.h>
 #include <string.h>
-#include <stdint.h>
 
 #define TRUE  1
 #define FALSE 0
@@ -190,8 +189,8 @@ struct TopModel_Struct{
   //double dbl_arr_test[3];
 
   /************** Model State **************/
-  char* serialized;
-  uint64_t serialized_length;
+  char* serialized;      /* serialized state buffer; NULL when none held */
+  int serialized_length; /* bytes in serialized; int because GetValue hands out its address */
 
 };
